Combine logging/{git,startup}.py
