+++ service_kit/logging/__init__.py
@@ -12,8 +12,8 @@ from .error_logging import (
     log_postgres_error as log_postgres_error,
     log_structured_error as log_structured_error,
 )
-from .git import log_git_status as log_git_status
 from .startup import (
+    log_git_status as log_git_status,
     log_python_version as log_python_version,
     log_startup_information as log_startup_information,
 )
